+++ lib/options/option_parser.rb
@@ -17,10 +17,12 @@ module Maze
           text 'Usage [OPTIONS] <filenames>'
           text ''
           text 'Overridden Cucumber options:'
-          opt :help, 'Print this help.'
-          opt :init, 'Initialises a new Maze Runner project'
-
-          opt :version, 'Display Maze Runner and Cucumber versions'
+          opt :help,
+              'Print this help.'
+          opt :init,
+              'Initialises a new Maze Runner project'
+          opt :version,
+              'Display Maze Runner and Cucumber versions'
 
           # Common options
           opt Option::SEPARATE_SESSIONS,
@@ -28,8 +30,14 @@ module Maze
               short: :none,
               type: :boolean,
               default: false
-          opt Option::FARM, 'Device farm to use: "bs" (BrowserStack) or "local"', short: '-f', type: :string
-          opt Option::APP, 'The app to be installed and run against', short: '-a', type: :string
+          opt Option::FARM,
+              'Device farm to use: "bs" (BrowserStack) or "local"',
+              short: '-f',
+              type: :string
+          opt Option::APP,
+              'The app to be installed and run against',
+              short: '-a',
+              type: :string
           opt Option::A11Y_LOCATOR,
               'Locate elements by accessibility id rather than id',
               short: :none,
@@ -46,16 +54,24 @@ module Maze
 
           # BrowserStack-only options
           opt Option::BS_LOCAL,
-              '(BS only) Path to the BrowserStackLocal binary',
+              '(BS only) Path to the BrowserStackLocal binary. MAZE_BS_LOCAL env var or "/BrowserStackLocal" by default',
               short: :none,
               type: :string,
-              default: '/BrowserStackLocal'
+              default: ENV['MAZE_BS_LOCAL'] || '/BrowserStackLocal'
           opt Option::BS_DEVICE,
               'BrowserStack device to use (a key of Devices.DEVICE_HASH)',
               short: :none,
               type: :string
-          opt Option::USERNAME, 'Device farm username', short: '-u', type: :string
-          opt Option::ACCESS_KEY, 'Device farm access key', short: '-p', type: :string
+          opt Option::USERNAME,
+              'Device farm username. MAZE_DEVICE_FARM_USERNAME env var by default',
+              short: '-u',
+              type: :string,
+              default: ENV['MAZE_DEVICE_FARM_USERNAME']
+          opt Option::ACCESS_KEY,
+              'Device farm access key. MAZE_DEVICE_FARM_ACCESS_KEY env var by default',
+              short: '-p',
+              type: :string,
+              default: ENV['MAZE_DEVICE_FARM_ACCESS_KEY']
           opt Option::BS_APPIUM_VERSION,
               'The Appium version to use with BrowserStack',
               short: :none,
@@ -71,12 +87,20 @@ module Maze
               short: :none,
               type: :string
           opt Option::APPIUM_SERVER,
-              'Appium server URL, only used for --farm=local',
+              'Appium server URL, only used for --farm=local. MAZE_APPIUM_SERVER env var by default',
               short: :none,
               type: :string,
-              default: 'http://localhost:4723/wd/hub'
-          opt Option::APPLE_TEAM_ID, 'Apple Team Id, required for local iOS testing', short: :none, type: :string
-          opt Option::UDID, 'Apple UDID, required for local iOS testing', short: :none, type: :string
+              default: ENV['MAZE_APPIUM_SERVER'] || 'http://localhost:4723/wd/hub'
+          opt Option::APPLE_TEAM_ID,
+              'Apple Team Id, required for local iOS testing. MAZE_APPLE_TEAM_ID env var by default',
+              short: :none,
+              type: :string,
+              default: ENV['MAZE_APPLE_TEAM_ID']
+          opt Option::UDID,
+              'Apple UDID, required for local iOS testing. MAZE_UDID env var by default',
+              short: :none,
+              type: :string,
+              default: ENV['MAZE_UDID']
 
           version "Maze Runner v#{BugsnagMazeRunner::VERSION} " \
                   "(Cucumber v#{Cucumber::VERSION.strip})"